Convert component files from TSX to JSX
