+++ src/components/Catering.jsx
@@ -1,201 +1,147 @@
-
 import React, { useState } from 'react';
 import { Card, CardContent, CardDescription, CardHeader, CardTitle, CardFooter } from '@/components/ui/card';
 import { Button } from '@/components/ui/button';
 import { Badge } from '@/components/ui/badge';
 import { CustomBadge } from '@/components/ui/custom-badge';
-import { 
-  Utensils, 
-  CalendarDays, 
-  PlusCircle, 
-  MoreVertical,
-  Clock,
-  Users,
-  Banknote,
-  ClipboardCheck,
-  Map,
-  Phone,
-  User,
-  Search,
-  Calendar,
-  ChevronRight
-} from 'lucide-react';
+import { Utensils, CalendarDays, PlusCircle, MoreVertical, Clock, Users, Banknote, ClipboardCheck, Map, Phone, User, Search, Calendar, ChevronRight } from 'lucide-react';
 import { Input } from '@/components/ui/input';
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs';
 import { DropdownMenu, DropdownMenuContent, DropdownMenuItem, DropdownMenuLabel, DropdownMenuSeparator, DropdownMenuTrigger } from '@/components/ui/dropdown-menu';
 import { Avatar, AvatarFallback } from '@/components/ui/avatar';
-
-interface CateringEvent {
-  id: string;
-  name: string;
-  client: string;
-  date: string;
-  time: string;
-  location: string;
-  attendees: number;
-  status: 'scheduled' | 'in-progress' | 'completed' | 'cancelled';
-  total: number;
-  contactPerson: {
-    name: string;
-    phone: string;
-  };
-}
-
-interface MenuItem {
-  id: string;
-  name: string;
-  category: string;
-  price: number;
-  description: string;
-  popular: boolean;
-}
-
-const Catering: React.FC = () => {
-  const [searchTerm, setSearchTerm] = useState('');
-  const [currentTab, setCurrentTab] = useState<string>('upcoming');
-  
-  const [events, setEvents] = useState<CateringEvent[]>([
-    {
-      id: '1',
-      name: 'Corporate Lunch Meeting',
-      client: 'ABC Technologies',
-      date: '2025-04-20',
-      time: '12:30 - 14:00',
-      location: 'ABC Technologies HQ, Conference Room B',
-      attendees: 25,
-      status: 'scheduled',
-      total: 625.00,
-      contactPerson: {
-        name: 'John Smith',
-        phone: '555-123-4567'
-      }
-    },
-    {
-      id: '2',
-      name: 'Executive Breakfast',
-      client: 'Global Finance',
-      date: '2025-04-21',
-      time: '08:00 - 09:30',
-      location: 'Global Finance Tower, 15th Floor',
-      attendees: 12,
-      status: 'scheduled',
-      total: 360.00,
-      contactPerson: {
-        name: 'Maria Garcia',
-        phone: '555-987-6543'
-      }
-    },
-    {
-      id: '3',
-      name: 'Team Building Lunch',
-      client: 'InnovateTech',
-      date: '2025-05-05',
-      time: '11:30 - 13:30',
-      location: 'City Park Pavilion',
-      attendees: 45,
-      status: 'scheduled',
-      total: 1125.00,
-      contactPerson: {
-        name: 'Alex Johnson',
-        phone: '555-456-7890'
-      }
-    },
-    {
-      id: '4',
-      name: 'Charity Gala Dinner',
-      client: 'Hope Foundation',
-      date: '2025-05-15',
-      time: '18:00 - 22:00',
-      location: 'Grand Hotel Ballroom',
-      attendees: 120,
-      status: 'scheduled',
-      total: 6000.00,
-      contactPerson: {
-        name: 'Sarah Williams',
-        phone: '555-789-0123'
-      }
-    }
-  ]);
-
-  const [cateringMenu, setCateringMenu] = useState<MenuItem[]>([
-    {
-      id: '1',
-      name: 'Gourmet Sandwich Platter',
-      category: 'Platters',
-      price: 75.00,
-      description: 'Assortment of premium sandwiches with artisan breads and fillings',
-      popular: true
-    },
-    {
-      id: '2',
-      name: 'Mediterranean Mezze',
-      category: 'Appetizers',
-      price: 65.00,
-      description: 'Hummus, tzatziki, baba ganoush, olives, and pita bread',
-      popular: true
-    },
-    {
-      id: '3',
-      name: 'Executive Hot Lunch',
-      category: 'Entrees',
-      price: 25.00,
-      description: 'Per person: Choice of protein, two sides, and dessert',
-      popular: true
-    },
-    {
-      id: '4',
-      name: 'Fresh Fruit Platter',
-      category: 'Platters',
-      price: 45.00,
-      description: 'Seasonal fruits arranged beautifully',
-      popular: false
-    },
-    {
-      id: '5',
-      name: 'Artisan Cheese Board',
-      category: 'Appetizers',
-      price: 85.00,
-      description: 'Selection of fine cheeses with crackers and accompaniments',
-      popular: true
-    },
-    {
-      id: '6',
-      name: 'Breakfast Package',
-      category: 'Breakfast',
-      price: 15.00,
-      description: 'Per person: Pastries, fruit, yogurt, and coffee',
-      popular: false
-    }
-  ]);
-  
-  const getInitials = (name: string) => {
-    return name.split(' ').map(n => n[0]).join('').toUpperCase();
-  };
-  
-  const getStatusBadgeVariant = (status: string) => {
-    switch (status) {
-      case 'scheduled': return 'outline';
-      case 'in-progress': return 'default';
-      case 'completed': return 'success';
-      case 'cancelled': return 'destructive';
-      default: return 'outline';
-    }
-  };
-
-  const filteredEvents = events.filter(event => {
-    const matchesSearch = 
-      event.name.toLowerCase().includes(searchTerm.toLowerCase()) ||
-      event.client.toLowerCase().includes(searchTerm.toLowerCase());
-    
-    return matchesSearch;
-  });
-  
-  const sortedEvents = [...filteredEvents].sort((a, b) => 
-    new Date(a.date).getTime() - new Date(b.date).getTime()
-  );
-
-  return (
-    <div className="grid gap-4 md:grid-cols-3">
+const Catering = () => {
+    const [searchTerm, setSearchTerm] = useState('');
+    const [currentTab, setCurrentTab] = useState('upcoming');
+    const [events, setEvents] = useState([
+        {
+            id: '1',
+            name: 'Corporate Lunch Meeting',
+            client: 'ABC Technologies',
+            date: '2025-04-20',
+            time: '12:30 - 14:00',
+            location: 'ABC Technologies HQ, Conference Room B',
+            attendees: 25,
+            status: 'scheduled',
+            total: 625.00,
+            contactPerson: {
+                name: 'John Smith',
+                phone: '555-123-4567'
+            }
+        },
+        {
+            id: '2',
+            name: 'Executive Breakfast',
+            client: 'Global Finance',
+            date: '2025-04-21',
+            time: '08:00 - 09:30',
+            location: 'Global Finance Tower, 15th Floor',
+            attendees: 12,
+            status: 'scheduled',
+            total: 360.00,
+            contactPerson: {
+                name: 'Maria Garcia',
+                phone: '555-987-6543'
+            }
+        },
+        {
+            id: '3',
+            name: 'Team Building Lunch',
+            client: 'InnovateTech',
+            date: '2025-05-05',
+            time: '11:30 - 13:30',
+            location: 'City Park Pavilion',
+            attendees: 45,
+            status: 'scheduled',
+            total: 1125.00,
+            contactPerson: {
+                name: 'Alex Johnson',
+                phone: '555-456-7890'
+            }
+        },
+        {
+            id: '4',
+            name: 'Charity Gala Dinner',
+            client: 'Hope Foundation',
+            date: '2025-05-15',
+            time: '18:00 - 22:00',
+            location: 'Grand Hotel Ballroom',
+            attendees: 120,
+            status: 'scheduled',
+            total: 6000.00,
+            contactPerson: {
+                name: 'Sarah Williams',
+                phone: '555-789-0123'
+            }
+        }
+    ]);
+    const [cateringMenu, setCateringMenu] = useState([
+        {
+            id: '1',
+            name: 'Gourmet Sandwich Platter',
+            category: 'Platters',
+            price: 75.00,
+            description: 'Assortment of premium sandwiches with artisan breads and fillings',
+            popular: true
+        },
+        {
+            id: '2',
+            name: 'Mediterranean Mezze',
+            category: 'Appetizers',
+            price: 65.00,
+            description: 'Hummus, tzatziki, baba ganoush, olives, and pita bread',
+            popular: true
+        },
+        {
+            id: '3',
+            name: 'Executive Hot Lunch',
+            category: 'Entrees',
+            price: 25.00,
+            description: 'Per person: Choice of protein, two sides, and dessert',
+            popular: true
+        },
+        {
+            id: '4',
+            name: 'Fresh Fruit Platter',
+            category: 'Platters',
+            price: 45.00,
+            description: 'Seasonal fruits arranged beautifully',
+            popular: false
+        },
+        {
+            id: '5',
+            name: 'Artisan Cheese Board',
+            category: 'Appetizers',
+            price: 85.00,
+            description: 'Selection of fine cheeses with crackers and accompaniments',
+            popular: true
+        },
+        {
+            id: '6',
+            name: 'Breakfast Package',
+            category: 'Breakfast',
+            price: 15.00,
+            description: 'Per person: Pastries, fruit, yogurt, and coffee',
+            popular: false
+        }
+    ]);
+    const getInitials = (name) => {
+        return name.split(' ').map(n => n[0]).join('').toUpperCase();
+    };
+    const getStatusBadgeVariant = (status) => {
+        switch (status) {
+            case 'scheduled': return 'outline';
+            case 'in-progress': return 'default';
+            case 'completed': return 'success';
+            case 'cancelled': return 'destructive';
+            default: return 'outline';
+        }
+    };
+    const filteredEvents = events.filter(event => {
+        const matchesSearch = event.name.toLowerCase().includes(searchTerm.toLowerCase()) ||
+            event.client.toLowerCase().includes(searchTerm.toLowerCase());
+        return matchesSearch;
+    });
+    const sortedEvents = [...filteredEvents].sort((a, b) => new Date(a.date).getTime() - new Date(b.date).getTime());
+    return (<div className="grid gap-4 md:grid-cols-3">
       <div className="md:col-span-2 space-y-4">
         <Card>
           <CardHeader className="flex flex-row items-center justify-between pb-2">
@@ -204,23 +150,17 @@ const Catering: React.FC = () => {
               <CardDescription>Handle catering orders and events</CardDescription>
             </div>
             <Button className="flex gap-1 items-center">
-              <PlusCircle className="h-4 w-4 mr-1" /> New Event
+              <PlusCircle className="h-4 w-4 mr-1"/> New Event
             </Button>
           </CardHeader>
           <CardContent className="space-y-4">
             <div className="flex flex-col space-y-2 md:flex-row md:space-x-2 md:space-y-0">
               <div className="relative flex-1">
-                <Search className="absolute left-2.5 top-2.5 h-4 w-4 text-muted-foreground" />
-                <Input
-                  type="search"
-                  placeholder="Search events..."
-                  className="pl-8"
-                  value={searchTerm}
-                  onChange={(e) => setSearchTerm(e.target.value)}
-                />
+                <Search className="absolute left-2.5 top-2.5 h-4 w-4 text-muted-foreground"/>
+                <Input type="search" placeholder="Search events..." className="pl-8" value={searchTerm} onChange={(e) => setSearchTerm(e.target.value)}/>
               </div>
               <Button variant="outline" className="flex gap-1 items-center">
-                <Calendar className="h-4 w-4 mr-1" /> Calendar View
+                <Calendar className="h-4 w-4 mr-1"/> Calendar View
               </Button>
             </div>
             
@@ -231,8 +171,7 @@ const Catering: React.FC = () => {
                 <TabsTrigger value="cancelled">Cancelled</TabsTrigger>
               </TabsList>
               <TabsContent value="upcoming" className="pt-2">
-                {sortedEvents.length > 0 ? (
-                  <div className="rounded-md border">
+                {sortedEvents.length > 0 ? (<div className="rounded-md border">
                     <div className="relative w-full overflow-auto">
                       <table className="w-full caption-bottom text-sm">
                         <thead>
@@ -245,8 +184,7 @@ const Catering: React.FC = () => {
                           </tr>
                         </thead>
                         <tbody>
-                          {sortedEvents.map((event) => (
-                            <tr key={event.id} className="border-b transition-colors hover:bg-muted/50">
+                          {sortedEvents.map((event) => (<tr key={event.id} className="border-b transition-colors hover:bg-muted/50">
                               <td className="p-4 align-middle font-medium">
                                 {event.name}
                               </td>
@@ -256,18 +194,18 @@ const Catering: React.FC = () => {
                               <td className="p-4 align-middle whitespace-nowrap">
                                 <div className="flex flex-col">
                                   <div className="flex items-center gap-1">
-                                    <CalendarDays className="h-3 w-3" />
+                                    <CalendarDays className="h-3 w-3"/>
                                     <span>{event.date}</span>
                                   </div>
                                   <div className="flex items-center gap-1 text-sm text-muted-foreground">
-                                    <Clock className="h-3 w-3" />
+                                    <Clock className="h-3 w-3"/>
                                     <span>{event.time}</span>
                                   </div>
                                 </div>
                               </td>
                               <td className="p-4 align-middle hidden md:table-cell">
                                 <div className="flex items-center gap-1">
-                                  <Users className="h-3 w-3" />
+                                  <Users className="h-3 w-3"/>
                                   <span>{event.attendees}</span>
                                 </div>
                               </td>
@@ -275,43 +213,39 @@ const Catering: React.FC = () => {
                                 <DropdownMenu>
                                   <DropdownMenuTrigger asChild>
                                     <Button variant="ghost" size="sm">
-                                      <MoreVertical className="h-4 w-4" />
+                                      <MoreVertical className="h-4 w-4"/>
                                     </Button>
                                   </DropdownMenuTrigger>
                                   <DropdownMenuContent align="end">
                                     <DropdownMenuLabel>Actions</DropdownMenuLabel>
                                     <DropdownMenuSeparator />
                                     <DropdownMenuItem>
-                                      <ClipboardCheck className="mr-2 h-4 w-4" /> View Details
+                                      <ClipboardCheck className="mr-2 h-4 w-4"/> View Details
                                     </DropdownMenuItem>
                                     <DropdownMenuItem>
-                                      <Users className="mr-2 h-4 w-4" /> Staff Assignment
+                                      <Users className="mr-2 h-4 w-4"/> Staff Assignment
                                     </DropdownMenuItem>
                                     <DropdownMenuItem>
-                                      <Utensils className="mr-2 h-4 w-4" /> Menu Items
+                                      <Utensils className="mr-2 h-4 w-4"/> Menu Items
                                     </DropdownMenuItem>
                                     <DropdownMenuSeparator />
                                     <DropdownMenuItem className="text-destructive">
-                                      <MoreVertical className="mr-2 h-4 w-4" /> Cancel Event
+                                      <MoreVertical className="mr-2 h-4 w-4"/> Cancel Event
                                     </DropdownMenuItem>
                                   </DropdownMenuContent>
                                 </DropdownMenu>
                               </td>
-                            </tr>
-                          ))}
+                            </tr>))}
                         </tbody>
                       </table>
                     </div>
-                  </div>
-                ) : (
-                  <div className="text-center py-10">
-                    <Utensils className="mx-auto h-12 w-12 text-muted-foreground/50 mb-3" />
+                  </div>) : (<div className="text-center py-10">
+                    <Utensils className="mx-auto h-12 w-12 text-muted-foreground/50 mb-3"/>
                     <p className="text-muted-foreground">No upcoming catering events found</p>
                     <Button className="mt-4" variant="outline" size="sm">
                       Create New Event
                     </Button>
-                  </div>
-                )}
+                  </div>)}
               </TabsContent>
               <TabsContent value="past" className="pt-2">
                 <div className="text-center py-10">
@@ -327,8 +261,7 @@ const Catering: React.FC = () => {
           </CardContent>
         </Card>
         
-        {currentTab === 'upcoming' && sortedEvents.length > 0 && (
-          <Card>
+        {currentTab === 'upcoming' && sortedEvents.length > 0 && (<Card>
             <CardHeader>
               <CardTitle>Event Details</CardTitle>
               <CardDescription>Next scheduled catering event</CardDescription>
@@ -348,7 +281,7 @@ const Catering: React.FC = () => {
                 <div className="grid grid-cols-1 md:grid-cols-3 gap-4">
                   <div className="flex flex-col p-3 border rounded-md">
                     <div className="flex items-center gap-2 text-sm font-medium mb-1">
-                      <CalendarDays className="h-4 w-4 text-primary" />
+                      <CalendarDays className="h-4 w-4 text-primary"/>
                       Date & Time
                     </div>
                     <p className="text-sm">
@@ -359,7 +292,7 @@ const Catering: React.FC = () => {
                   
                   <div className="flex flex-col p-3 border rounded-md">
                     <div className="flex items-center gap-2 text-sm font-medium mb-1">
-                      <Map className="h-4 w-4 text-primary" />
+                      <Map className="h-4 w-4 text-primary"/>
                       Location
                     </div>
                     <p className="text-sm">{sortedEvents[0].location}</p>
@@ -367,7 +300,7 @@ const Catering: React.FC = () => {
                   
                   <div className="flex flex-col p-3 border rounded-md">
                     <div className="flex items-center gap-2 text-sm font-medium mb-1">
-                      <Users className="h-4 w-4 text-primary" />
+                      <Users className="h-4 w-4 text-primary"/>
                       Attendees
                     </div>
                     <p className="text-sm">{sortedEvents[0].attendees} people</p>
@@ -377,7 +310,7 @@ const Catering: React.FC = () => {
                 <div className="flex flex-col md:flex-row gap-4">
                   <div className="flex-1 p-3 border rounded-md">
                     <div className="flex items-center gap-2 text-sm font-medium mb-2">
-                      <User className="h-4 w-4 text-primary" />
+                      <User className="h-4 w-4 text-primary"/>
                       Contact Person
                     </div>
                     <div className="flex items-center gap-3">
@@ -387,7 +320,7 @@ const Catering: React.FC = () => {
                       <div>
                         <p className="font-medium">{sortedEvents[0].contactPerson.name}</p>
                         <div className="flex items-center gap-1 text-sm text-muted-foreground">
-                          <Phone className="h-3 w-3" />
+                          <Phone className="h-3 w-3"/>
                           <span>{sortedEvents[0].contactPerson.phone}</span>
                         </div>
                       </div>
@@ -396,7 +329,7 @@ const Catering: React.FC = () => {
                   
                   <div className="flex-1 p-3 border rounded-md">
                     <div className="flex items-center gap-2 text-sm font-medium mb-2">
-                      <Banknote className="h-4 w-4 text-primary" />
+                      <Banknote className="h-4 w-4 text-primary"/>
                       Financial
                     </div>
                     <div className="space-y-1">
@@ -420,11 +353,10 @@ const Catering: React.FC = () => {
             <CardFooter className="flex justify-between border-t pt-4">
               <Button variant="outline" size="sm">View Full Details</Button>
               <Button variant="default" size="sm">
-                Generate Report <ChevronRight className="ml-1 h-4 w-4" />
+                Generate Report <ChevronRight className="ml-1 h-4 w-4"/>
               </Button>
             </CardFooter>
-          </Card>
-        )}
+          </Card>)}
       </div>
       
       <div className="space-y-4">
@@ -436,24 +368,20 @@ const Catering: React.FC = () => {
           <CardContent>
             <div className="space-y-4">
               {cateringMenu
-                .filter((item, index) => index < 4)
-                .map((item) => (
-                <div key={item.id} className="flex justify-between items-start border-b pb-3 last:border-0 last:pb-0">
+            .filter((item, index) => index < 4)
+            .map((item) => (<div key={item.id} className="flex justify-between items-start border-b pb-3 last:border-0 last:pb-0">
                   <div>
                     <p className="font-medium flex items-center">
                       {item.name}
-                      {item.popular && (
-                        <Badge variant="secondary" className="ml-2 text-xs">Popular</Badge>
-                      )}
+                      {item.popular && (<Badge variant="secondary" className="ml-2 text-xs">Popular</Badge>)}
                     </p>
                     <p className="text-xs text-muted-foreground mt-1">{item.description}</p>
                     <p className="text-sm mt-1">${item.price.toFixed(2)}</p>
                   </div>
                   <Button variant="ghost" size="sm">
-                    <MoreVertical className="h-4 w-4" />
+                    <MoreVertical className="h-4 w-4"/>
                   </Button>
-                </div>
-              ))}
+                </div>))}
             </div>
             <Button className="w-full mt-4" variant="outline" size="sm">
               View Full Menu ({cateringMenu.length} items)
@@ -467,12 +395,10 @@ const Catering: React.FC = () => {
             <CardDescription>Next 30 days</CardDescription>
           </CardHeader>
           <CardContent>
-            {events.length > 0 ? (
-              <div className="space-y-3">
+            {events.length > 0 ? (<div className="space-y-3">
                 {events
-                  .filter((_, index) => index < 4)
-                  .map((event) => (
-                  <div key={event.id} className="flex items-center gap-3 border-b pb-3 last:border-0 last:pb-0">
+                .filter((_, index) => index < 4)
+                .map((event) => (<div key={event.id} className="flex items-center gap-3 border-b pb-3 last:border-0 last:pb-0">
                     <div className="flex-shrink-0 flex flex-col items-center justify-center w-12 h-12 rounded-md bg-primary/10 text-primary">
                       <span className="text-sm font-bold">
                         {event.date.split('-')[2]}
@@ -485,24 +411,18 @@ const Catering: React.FC = () => {
                       <p className="font-medium truncate">{event.name}</p>
                       <p className="text-xs text-muted-foreground">{event.client}</p>
                       <div className="flex items-center gap-1 text-xs text-muted-foreground mt-1">
-                        <Clock className="h-3 w-3" />
+                        <Clock className="h-3 w-3"/>
                         <span>{event.time}</span>
                       </div>
                     </div>
-                  </div>
-                ))}
-              </div>
-            ) : (
-              <div className="text-center py-6">
-                <CalendarDays className="mx-auto h-12 w-12 text-muted-foreground/50 mb-3" />
+                  </div>))}
+              </div>) : (<div className="text-center py-6">
+                <CalendarDays className="mx-auto h-12 w-12 text-muted-foreground/50 mb-3"/>
                 <p className="text-muted-foreground">No upcoming events scheduled</p>
-              </div>
-            )}
+              </div>)}
           </CardContent>
         </Card>
       </div>
-    </div>
-  );
+    </div>);
 };
-
 export default Catering;
