+++ src/components/Dashboard.jsx
@@ -1,24 +1,19 @@
-
 import React from 'react';
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card';
-import { dashboardStats, salesData, menuItems } from '@/utils/mockData';
+import { dashboardStats, salesData } from '@/utils/mockData';
 import { AreaChart, Area, BarChart, Bar, XAxis, YAxis, CartesianGrid, Tooltip, Legend, ResponsiveContainer } from 'recharts';
-import { TrendingUp, TrendingDown, Users, ShoppingBag, DollarSign } from 'lucide-react';
-
-const Dashboard: React.FC = () => {
-  // Format data for charts
-  const salesTimeData = dashboardStats.salesByTime.map(item => ({
-    name: item.time,
-    amount: item.amount
-  }));
-  
-  const categorySalesData = dashboardStats.salesByCategory.map(item => ({
-    name: item.category,
-    amount: item.amount
-  }));
-
-  return (
-    <div className="space-y-6 animate-fade-in">
+import { TrendingUp, Users, ShoppingBag, DollarSign } from 'lucide-react';
+const Dashboard = () => {
+    // Format data for charts
+    const salesTimeData = dashboardStats.salesByTime.map(item => ({
+        name: item.time,
+        amount: item.amount
+    }));
+    const categorySalesData = dashboardStats.salesByCategory.map(item => ({
+        name: item.category,
+        amount: item.amount
+    }));
+    return (<div className="space-y-6 animate-fade-in">
       <h2 className="text-3xl font-semibold">Dashboard</h2>
       
       {/* Stats Cards Row */}
@@ -26,7 +21,7 @@ const Dashboard: React.FC = () => {
         <Card>
           <CardHeader className="flex flex-row items-center justify-between pb-2">
             <CardTitle className="text-sm font-medium">Today's Sales</CardTitle>
-            <DollarSign className="h-4 w-4 text-muted-foreground" />
+            <DollarSign className="h-4 w-4 text-muted-foreground"/>
           </CardHeader>
           <CardContent>
             <div className="text-2xl font-bold">₱{dashboardStats.dailySales.toFixed(2)}</div>
@@ -39,7 +34,7 @@ const Dashboard: React.FC = () => {
         <Card>
           <CardHeader className="flex flex-row items-center justify-between pb-2">
             <CardTitle className="text-sm font-medium">Monthly Sales</CardTitle>
-            <TrendingUp className="h-4 w-4 text-muted-foreground" />
+            <TrendingUp className="h-4 w-4 text-muted-foreground"/>
           </CardHeader>
           <CardContent>
             <div className="text-2xl font-bold">₱{dashboardStats.monthlySales.toFixed(2)}</div>
@@ -52,7 +47,7 @@ const Dashboard: React.FC = () => {
         <Card>
           <CardHeader className="flex flex-row items-center justify-between pb-2">
             <CardTitle className="text-sm font-medium">Customers Today</CardTitle>
-            <Users className="h-4 w-4 text-muted-foreground" />
+            <Users className="h-4 w-4 text-muted-foreground"/>
           </CardHeader>
           <CardContent>
             <div className="text-2xl font-bold">{dashboardStats.customerCount}</div>
@@ -65,7 +60,7 @@ const Dashboard: React.FC = () => {
         <Card>
           <CardHeader className="flex flex-row items-center justify-between pb-2">
             <CardTitle className="text-sm font-medium">Orders Today</CardTitle>
-            <ShoppingBag className="h-4 w-4 text-muted-foreground" />
+            <ShoppingBag className="h-4 w-4 text-muted-foreground"/>
           </CardHeader>
           <CardContent>
             <div className="text-2xl font-bold">{salesData.length}</div>
@@ -87,19 +82,18 @@ const Dashboard: React.FC = () => {
           </CardHeader>
           <CardContent className="h-80">
             <ResponsiveContainer width="100%" height="100%">
-              <AreaChart data={salesTimeData}
-                margin={{ top: 10, right: 30, left: 0, bottom: 0 }}>
+              <AreaChart data={salesTimeData} margin={{ top: 10, right: 30, left: 0, bottom: 0 }}>
                 <defs>
                   <linearGradient id="colorUv" x1="0" y1="0" x2="0" y2="1">
                     <stop offset="5%" stopColor="hsl(var(--primary))" stopOpacity={0.8}/>
                     <stop offset="95%" stopColor="hsl(var(--primary))" stopOpacity={0.1}/>
                   </linearGradient>
                 </defs>
-                <XAxis dataKey="name" />
+                <XAxis dataKey="name"/>
                 <YAxis />
-                <CartesianGrid strokeDasharray="3 3" />
+                <CartesianGrid strokeDasharray="3 3"/>
                 <Tooltip />
-                <Area type="monotone" dataKey="amount" stroke="hsl(var(--primary))" fillOpacity={1} fill="url(#colorUv)" />
+                <Area type="monotone" dataKey="amount" stroke="hsl(var(--primary))" fillOpacity={1} fill="url(#colorUv)"/>
               </AreaChart>
             </ResponsiveContainer>
           </CardContent>
@@ -115,12 +109,12 @@ const Dashboard: React.FC = () => {
           <CardContent className="h-80">
             <ResponsiveContainer width="100%" height="100%">
               <BarChart data={categorySalesData}>
-                <CartesianGrid strokeDasharray="3 3" />
-                <XAxis dataKey="name" />
+                <CartesianGrid strokeDasharray="3 3"/>
+                <XAxis dataKey="name"/>
                 <YAxis />
                 <Tooltip />
                 <Legend />
-                <Bar dataKey="amount" fill="hsl(var(--secondary))" />
+                <Bar dataKey="amount" fill="hsl(var(--secondary))"/>
               </BarChart>
             </ResponsiveContainer>
           </CardContent>
@@ -138,15 +132,13 @@ const Dashboard: React.FC = () => {
           </CardHeader>
           <CardContent>
             <div className="space-y-4">
-              {dashboardStats.popularItems.map((item, index) => (
-                <div key={index} className="flex items-center justify-between">
+              {dashboardStats.popularItems.map((item, index) => (<div key={index} className="flex items-center justify-between">
                   <div className="flex items-center space-x-4">
                     <span className="font-medium">{index + 1}.</span>
                     <span>{item.name}</span>
                   </div>
                   <span className="text-muted-foreground">{item.count} orders</span>
-                </div>
-              ))}
+                </div>))}
             </div>
           </CardContent>
         </Card>
@@ -160,8 +152,7 @@ const Dashboard: React.FC = () => {
           </CardHeader>
           <CardContent>
             <div className="space-y-4">
-              {dashboardStats.recentSales.map((sale) => (
-                <div key={sale.id} className="flex items-center justify-between">
+              {dashboardStats.recentSales.map((sale) => (<div key={sale.id} className="flex items-center justify-between">
                   <div>
                     <p className="font-medium">Order #{sale.id}</p>
                     <p className="text-sm text-muted-foreground">
@@ -172,14 +163,11 @@ const Dashboard: React.FC = () => {
                     <p className="font-medium">₱{sale.total.toFixed(2)}</p>
                     <p className="text-sm text-muted-foreground capitalize">{sale.paymentMethod}</p>
                   </div>
-                </div>
-              ))}
+                </div>))}
             </div>
           </CardContent>
         </Card>
       </div>
-    </div>
-  );
+    </div>);
 };
-
 export default Dashboard;
