+++ src/components/MenuManagement.jsx
@@ -8,77 +8,64 @@ import { Dialog, DialogContent, DialogDescription, DialogFooter, DialogHeader, D
 import { Switch } from '@/components/ui/switch';
 import { Badge } from '@/components/ui/badge';
 import { menuItems } from '@/utils/mockData';
-import { MenuItem } from '@/types';
-import { Edit, Plus, Trash2, Eye } from 'lucide-react';
+import { Edit, Plus, Trash2 } from 'lucide-react';
 import { toast } from 'sonner';
-
-const MenuManagement: React.FC = () => {
-  const [items, setItems] = useState<MenuItem[]>(menuItems);
-  const [newItem, setNewItem] = useState<Partial<MenuItem>>({
-    name: '',
-    description: '',
-    price: 0,
-    category: '',
-    available: true,
-    popular: false
-  });
-  const [dialogOpen, setDialogOpen] = useState(false);
-  const [editingItem, setEditingItem] = useState<MenuItem | null>(null);
-  
-  const categories = Array.from(new Set(items.map(item => item.category)));
-  
-  const handleAddItem = () => {
-    if (!newItem.name || !newItem.description || !newItem.category) {
-      toast.error('Please fill in all required fields');
-      return;
-    }
-    
-    const itemToAdd = {
-      ...newItem,
-      id: `P${items.length + 1}`,
-      price: Number(newItem.price),
-      available: newItem.available ?? true,
-      popular: newItem.popular ?? false
-    } as MenuItem;
-
-    setItems([...items, itemToAdd]);
-    setNewItem({
-      name: '',
-      description: '',
-      price: 0,
-      category: '',
-      available: true,
-      popular: false
+const MenuManagement = () => {
+    const [items, setItems] = useState(menuItems);
+    const [newItem, setNewItem] = useState({
+        name: '',
+        description: '',
+        price: 0,
+        category: '',
+        available: true,
+        popular: false
     });
-    setDialogOpen(false);
-    toast.success('Menu item added successfully');
-  };
-
-  const handleEditItem = () => {
-    if (!editingItem) return;
-    
-    const updatedItems = items.map(item => 
-      item.id === editingItem.id ? editingItem : item
-    );
-    
-    setItems(updatedItems);
-    setEditingItem(null);
-    toast.success('Menu item updated successfully');
-  };
-
-  const handleDeleteItem = (id: string) => {
-    setItems(items.filter(item => item.id !== id));
-    toast.success('Menu item deleted successfully');
-  };
-
-  return (
-    <div className="space-y-6 animate-fade-in">
+    const [dialogOpen, setDialogOpen] = useState(false);
+    const [editingItem, setEditingItem] = useState(null);
+    const categories = Array.from(new Set(items.map(item => item.category)));
+    const handleAddItem = () => {
+        if (!newItem.name || !newItem.description || !newItem.category) {
+            toast.error('Please fill in all required fields');
+            return;
+        }
+        const itemToAdd = {
+            ...newItem,
+            id: `P${items.length + 1}`,
+            price: Number(newItem.price),
+            available: newItem.available ?? true,
+            popular: newItem.popular ?? false
+        };
+        setItems([...items, itemToAdd]);
+        setNewItem({
+            name: '',
+            description: '',
+            price: 0,
+            category: '',
+            available: true,
+            popular: false
+        });
+        setDialogOpen(false);
+        toast.success('Menu item added successfully');
+    };
+    const handleEditItem = () => {
+        if (!editingItem)
+            return;
+        const updatedItems = items.map(item => item.id === editingItem.id ? editingItem : item);
+        setItems(updatedItems);
+        setEditingItem(null);
+        toast.success('Menu item updated successfully');
+    };
+    const handleDeleteItem = (id) => {
+        setItems(items.filter(item => item.id !== id));
+        toast.success('Menu item deleted successfully');
+    };
+    return (<div className="space-y-6 animate-fade-in">
       <div className="flex items-center justify-between">
         <h2 className="text-3xl font-semibold">Menu Management</h2>
         <Dialog open={dialogOpen} onOpenChange={setDialogOpen}>
           <DialogTrigger asChild>
             <Button className="flex items-center gap-2">
-              <Plus size={16} /> Add Item
+              <Plus size={16}/> Add Item
             </Button>
           </DialogTrigger>
           <DialogContent>
@@ -91,56 +78,27 @@ const MenuManagement: React.FC = () => {
             <div className="grid gap-4 py-4">
               <div className="grid grid-cols-4 items-center gap-4">
                 <Label htmlFor="name" className="text-right">Name</Label>
-                <Input 
-                  id="name" 
-                  value={newItem.name} 
-                  onChange={(e) => setNewItem({...newItem, name: e.target.value})}
-                  className="col-span-3"
-                />
+                <Input id="name" value={newItem.name} onChange={(e) => setNewItem({ ...newItem, name: e.target.value })} className="col-span-3"/>
               </div>
               <div className="grid grid-cols-4 items-center gap-4">
                 <Label htmlFor="description" className="text-right">Description</Label>
-                <Input 
-                  id="description" 
-                  value={newItem.description} 
-                  onChange={(e) => setNewItem({...newItem, description: e.target.value})}
-                  className="col-span-3"
-                />
+                <Input id="description" value={newItem.description} onChange={(e) => setNewItem({ ...newItem, description: e.target.value })} className="col-span-3"/>
               </div>
               <div className="grid grid-cols-4 items-center gap-4">
                 <Label htmlFor="price" className="text-right">Price</Label>
-                <Input 
-                  id="price" 
-                  type="number" 
-                  value={newItem.price} 
-                  onChange={(e) => setNewItem({...newItem, price: parseFloat(e.target.value)})}
-                  className="col-span-3"
-                />
+                <Input id="price" type="number" value={newItem.price} onChange={(e) => setNewItem({ ...newItem, price: parseFloat(e.target.value) })} className="col-span-3"/>
               </div>
               <div className="grid grid-cols-4 items-center gap-4">
                 <Label htmlFor="category" className="text-right">Category</Label>
-                <Input 
-                  id="category" 
-                  value={newItem.category} 
-                  onChange={(e) => setNewItem({...newItem, category: e.target.value})}
-                  className="col-span-3"
-                />
+                <Input id="category" value={newItem.category} onChange={(e) => setNewItem({ ...newItem, category: e.target.value })} className="col-span-3"/>
               </div>
               <div className="grid grid-cols-4 items-center gap-4">
                 <Label htmlFor="available" className="text-right">Available</Label>
-                <Switch 
-                  id="available" 
-                  checked={newItem.available} 
-                  onCheckedChange={(checked) => setNewItem({...newItem, available: checked})}
-                />
+                <Switch id="available" checked={newItem.available} onCheckedChange={(checked) => setNewItem({ ...newItem, available: checked })}/>
               </div>
               <div className="grid grid-cols-4 items-center gap-4">
                 <Label htmlFor="popular" className="text-right">Popular</Label>
-                <Switch 
-                  id="popular" 
-                  checked={newItem.popular} 
-                  onCheckedChange={(checked) => setNewItem({...newItem, popular: checked})}
-                />
+                <Switch id="popular" checked={newItem.popular} onCheckedChange={(checked) => setNewItem({ ...newItem, popular: checked })}/>
               </div>
             </div>
             <DialogFooter>
@@ -154,15 +112,12 @@ const MenuManagement: React.FC = () => {
       <Tabs defaultValue="all" className="w-full">
         <TabsList>
           <TabsTrigger value="all">All Items</TabsTrigger>
-          {categories.map(category => (
-            <TabsTrigger key={category} value={category}>{category}</TabsTrigger>
-          ))}
+          {categories.map(category => (<TabsTrigger key={category} value={category}>{category}</TabsTrigger>))}
         </TabsList>
         
         <TabsContent value="all" className="mt-6">
           <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-6">
-            {items.map(item => (
-              <Card key={item.id}>
+            {items.map(item => (<Card key={item.id}>
                 <CardHeader>
                   <div className="flex justify-between">
                     <CardTitle>{item.name}</CardTitle>
@@ -181,22 +136,19 @@ const MenuManagement: React.FC = () => {
                 </CardContent>
                 <CardFooter className="flex justify-end space-x-2">
                   <Button variant="outline" size="sm" onClick={() => setEditingItem(item)}>
-                    <Edit className="h-4 w-4 mr-1" /> Edit
+                    <Edit className="h-4 w-4 mr-1"/> Edit
                   </Button>
                   <Button variant="destructive" size="sm" onClick={() => handleDeleteItem(item.id)}>
-                    <Trash2 className="h-4 w-4 mr-1" /> Delete
+                    <Trash2 className="h-4 w-4 mr-1"/> Delete
                   </Button>
                 </CardFooter>
-              </Card>
-            ))}
+              </Card>))}
           </div>
         </TabsContent>
         
-        {categories.map(category => (
-          <TabsContent key={category} value={category} className="mt-6">
+        {categories.map(category => (<TabsContent key={category} value={category} className="mt-6">
             <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-6">
-              {items.filter(item => item.category === category).map(item => (
-                <Card key={item.id}>
+              {items.filter(item => item.category === category).map(item => (<Card key={item.id}>
                   <CardHeader>
                     <div className="flex justify-between">
                       <CardTitle>{item.name}</CardTitle>
@@ -214,22 +166,19 @@ const MenuManagement: React.FC = () => {
                   </CardContent>
                   <CardFooter className="flex justify-end space-x-2">
                     <Button variant="outline" size="sm" onClick={() => setEditingItem(item)}>
-                      <Edit className="h-4 w-4 mr-1" /> Edit
+                      <Edit className="h-4 w-4 mr-1"/> Edit
                     </Button>
                     <Button variant="destructive" size="sm" onClick={() => handleDeleteItem(item.id)}>
-                      <Trash2 className="h-4 w-4 mr-1" /> Delete
+                      <Trash2 className="h-4 w-4 mr-1"/> Delete
                     </Button>
                   </CardFooter>
-                </Card>
-              ))}
+                </Card>))}
             </div>
-          </TabsContent>
-        ))}
+          </TabsContent>))}
       </Tabs>
       
       {/* Edit Dialog */}
-      {editingItem && (
-        <Dialog open={!!editingItem} onOpenChange={(open) => !open && setEditingItem(null)}>
+      {editingItem && (<Dialog open={!!editingItem} onOpenChange={(open) => !open && setEditingItem(null)}>
           <DialogContent>
             <DialogHeader>
               <DialogTitle>Edit Menu Item</DialogTitle>
@@ -240,56 +189,27 @@ const MenuManagement: React.FC = () => {
             <div className="grid gap-4 py-4">
               <div className="grid grid-cols-4 items-center gap-4">
                 <Label htmlFor="edit-name" className="text-right">Name</Label>
-                <Input 
-                  id="edit-name" 
-                  value={editingItem.name} 
-                  onChange={(e) => setEditingItem({...editingItem, name: e.target.value})}
-                  className="col-span-3"
-                />
+                <Input id="edit-name" value={editingItem.name} onChange={(e) => setEditingItem({ ...editingItem, name: e.target.value })} className="col-span-3"/>
               </div>
               <div className="grid grid-cols-4 items-center gap-4">
                 <Label htmlFor="edit-description" className="text-right">Description</Label>
-                <Input 
-                  id="edit-description" 
-                  value={editingItem.description} 
-                  onChange={(e) => setEditingItem({...editingItem, description: e.target.value})}
-                  className="col-span-3"
-                />
+                <Input id="edit-description" value={editingItem.description} onChange={(e) => setEditingItem({ ...editingItem, description: e.target.value })} className="col-span-3"/>
               </div>
               <div className="grid grid-cols-4 items-center gap-4">
                 <Label htmlFor="edit-price" className="text-right">Price</Label>
-                <Input 
-                  id="edit-price" 
-                  type="number" 
-                  value={editingItem.price} 
-                  onChange={(e) => setEditingItem({...editingItem, price: parseFloat(e.target.value)})}
-                  className="col-span-3"
-                />
+                <Input id="edit-price" type="number" value={editingItem.price} onChange={(e) => setEditingItem({ ...editingItem, price: parseFloat(e.target.value) })} className="col-span-3"/>
               </div>
               <div className="grid grid-cols-4 items-center gap-4">
                 <Label htmlFor="edit-category" className="text-right">Category</Label>
-                <Input 
-                  id="edit-category" 
-                  value={editingItem.category} 
-                  onChange={(e) => setEditingItem({...editingItem, category: e.target.value})}
-                  className="col-span-3"
-                />
+                <Input id="edit-category" value={editingItem.category} onChange={(e) => setEditingItem({ ...editingItem, category: e.target.value })} className="col-span-3"/>
               </div>
               <div className="grid grid-cols-4 items-center gap-4">
                 <Label htmlFor="edit-available" className="text-right">Available</Label>
-                <Switch 
-                  id="edit-available" 
-                  checked={editingItem.available} 
-                  onCheckedChange={(checked) => setEditingItem({...editingItem, available: checked})}
-                />
+                <Switch id="edit-available" checked={editingItem.available} onCheckedChange={(checked) => setEditingItem({ ...editingItem, available: checked })}/>
               </div>
               <div className="grid grid-cols-4 items-center gap-4">
                 <Label htmlFor="edit-popular" className="text-right">Popular</Label>
-                <Switch 
-                  id="edit-popular" 
-                  checked={editingItem.popular} 
-                  onCheckedChange={(checked) => setEditingItem({...editingItem, popular: checked})}
-                />
+                <Switch id="edit-popular" checked={editingItem.popular} onCheckedChange={(checked) => setEditingItem({ ...editingItem, popular: checked })}/>
               </div>
             </div>
             <DialogFooter>
@@ -297,10 +217,7 @@ const MenuManagement: React.FC = () => {
               <Button onClick={handleEditItem}>Save Changes</Button>
             </DialogFooter>
           </DialogContent>
-        </Dialog>
-      )}
-    </div>
-  );
+        </Dialog>)}
+    </div>);
 };
-
 export default MenuManagement;
